+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
-* Fix Windows pthreads being added to non-Windows platforms
+* Fix Windows pthreads being added to non-Windows platforms when cross compiling
 
 ## [0.21.1] - 2025-09-24